backend/requirements.txt
asgiref==3.7.2
Django==4.2
sqlparse==0.4.4
typing_extensions==4.8.0
Pillow==10.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
django-filter